requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2
python-dotenv==1.0.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1